Move third-party js not managed through npm/bower into a `js/libs` directory
This allows the code coverage runner to exclude them from the reports
